local database is now fully functionnal

+++ app/src/main/java/com/openclassrooms/realestatemanager/ui/activity/MainActivity.kt
@@ -105,13 +105,13 @@ class MainActivity : BaseActivity<ActivityMainBinding>() {
             ActivityCompat.startActivity(activity, intent, null)
         }
 
-        fun navigateToAddEditActivity(activity: FragmentActivity, estateId: Long) {
+        fun navigateToAddEditActivity(activity: FragmentActivity, estateId: Int) {
             val intent = Intent(activity, AddEditEstateActivity::class.java)
             intent.putExtra("estate_id", estateId)
             ActivityCompat.startActivity(activity, intent, null)
         }
 
-        fun navigateToEstateDetailsActivity(activity: FragmentActivity, estateId: Long) {
+        fun navigateToEstateDetailsActivity(activity: FragmentActivity, estateId: Int) {
             val intent = Intent(activity, EstateDetailsActivity::class.java)
             intent.putExtra("estate_id", estateId)
             ActivityCompat.startActivity(activity, intent, null)
